refactor: rename constant.ts to constants.ts
Solves issue #22 point 2
